COMP: get rid of CLang compiler warnings
git-svn-id: https://advants.svn.sourceforge.net/svnroot/advants@1152 fd2da0a4-d968-48df-9cb5-d8c496287cbf

+++ Examples/WarpImageMultiTransform.cxx
@@ -209,7 +209,7 @@ bool ParseInput(int argc, char **argv, char *&moving_image_filename,
             if(strchr(s, 'x'))
               {
               char *tok = strtok(s, "x");
-              for(size_t i = 0; i < NDimensions; i++)
+              for(int i = 0; i < NDimensions; i++)
                 {
                 if(tok == NULL)
                   std::cerr << "Invalid sigma specification:" << s << std::endl;
